pyproject.toml
[project]
name = "one-stream-frame"
version = "0.2.0"
description = ""
authors = [
    {name = "Paul Bailey", email = "paul@neutron.studio"},
]
dependencies = [
    "google-auth>=2.6.0",
    "google-auth-httplib2>=0.1.0",
    "google-api-python-client>=2.39.0",
    "google-auth-oauthlib>=0.5.0",
]
requires-python = ">=3.9"
license = {text = "MIT"}

[project.urls]
homepage = ""

[tool]
[tool.pdm]

[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"
